Add node modules
This commit starts the process for adding the backend for blog posting
functionality. It's on a sepreate branch from 'Main' because it is
currently using Github Pages to host the static version of the site,
while I figure out how/where I'd like to host the full version.

The plan (for now) is to have a simple blog tab that fetches posts from
the backend in reverse chronological order and displays the latest three.
An archive grouped by month will be linked at the bottom as well as being
shown in an aside div in the right column. Tags are a soft maybe at this
stage and comments are a hard no.