Switch to Hugo (#1) (#2)
* Add node modules

This commit starts the process for adding the backend for blog posting
functionality. It's on a sepreate branch from 'Main' because it is
currently using Github Pages to host the static version of the site,
while I figure out how/where I'd like to host the full version.

The plan (for now) is to have a simple blog tab that fetches posts from
the backend in reverse chronological order and displays the latest three.
An archive grouped by month will be linked at the bottom as well as being
shown in an aside div in the right column. Tags are a soft maybe at this
stage and comments are a hard no.

* Organize directory for express

* Add all pages to server directory

* Add .gitignore

* Add pug

* Untrack node_modules folder

* Add database

* Add create post page

* Integrate blog database

* Fix blog linking

* Add 'Date Created' to posts

* Fix 'Created Date'

The 'Created Date' was not updating correctly with new Date() so I
changed it to Date.now.

* Change header to flex-box

* Fix margin under navbar

* Style the blog tab

* Reverse post list

* Start from scratch

* install hugo

* Create new blank theme

* Start hugo server for the first time

* Add new theme to config

* Create bare minimum for template

* Create bare minimum for template

* Set basic metadata

* Add simple header

* Change header to use Hugo Menus

* Enable GitInfo

* Fix syntax error

* Wipe all

* Updgrade Hugo

* Move menus to theme config

* Enable blockquote alerts

* Add bootstrap

* Fix header

* Create pages for certs and projects

* Add Resume basics

* Cleanup layouts for certs and projects

* Fix site subtitle

* Add certs and projects

* Turn cert and project list into cards

* Remove button, make title the link

* Make all images no wider than their parent

* Use Bootstrap's padding class

* Add all current certs

* Update sitemap

* Add gutters

* Update index

* Group pages by type

* Set the main width to 650

* Typography tweaks

* Fix navbar

* Fix active link class

* Add thumbnail

* Small formatting fixes

* Minor style tweaks

* draft post

* Ignore Obsidian meta-data

* Remove Obsidian meta-data

* Make heading typography smaller

* Publish post